+++ src/Component/Result/Stats/Result.php
@@ -14,19 +14,14 @@ namespace Solarium\Component\Result\Stats;
  */
 class Result
 {
+    use ResultTrait;
+
     /**
      * Field name.
      *
      * @var string
      */
     protected $field;
-
-    /**
-     * Stats data.
-     *
-     * @var array
-     */
-    protected $stats;
 
     /**
      * Constructor.
@@ -51,114 +46,26 @@ class Result
     }
 
     /**
-     * Get min value.
-     *
-     * @return string|null
-     */
-    public function getMin(): ?string
-    {
-        return $this->getValue('min');
-    }
-
-    /**
-     * Get max value.
-     *
-     * @return string|null
-     */
-    public function getMax(): ?string
-    {
-        return $this->getValue('max');
-    }
-
-    /**
-     * Get sum value.
-     *
-     * @return string|null
-     */
-    public function getSum(): ?string
-    {
-        return $this->getValue('sum');
-    }
-
-    /**
-     * Get count value.
-     *
-     * @return int|null
-     */
-    public function getCount(): ?int
-    {
-        return (int) $this->getValue('count');
-    }
-
-    /**
-     * Get missing value.
-     *
-     * @return string|null
-     */
-    public function getMissing(): ?string
-    {
-        return $this->getValue('missing');
-    }
-
-    /**
-     * Get sumOfSquares value.
-     *
-     * @return string|null
-     */
-    public function getSumOfSquares(): ?string
-    {
-        return $this->getValue('sumOfSquares');
-    }
-
-    /**
-     * Get mean value.
-     *
-     * @return string|null
-     */
-    public function getMean(): ?string
-    {
-        return $this->getValue('mean');
-    }
-
-    /**
-     * Get stddev value.
-     *
-     * @return string|null
-     */
-    public function getStddev(): ?string
-    {
-        return $this->getValue('stddev');
-    }
-
-    /**
      * Get facet stats.
      *
      * @return array|null
      */
     public function getFacets(): ?array
     {
-        return $this->getValue('facets');
+        return $this->getStatValue('facets');
     }
 
     /**
-     * Get percentile stats.
+     * Get value by stat name.
      *
-     * @return array|null
+     * @param string $stat
+     *
+     * @return mixed|null
+     *
+     * @deprecated Use getStatValue() instead
      */
-    public function getPercentiles(): ?array
+    public function getValue(string $stat)
     {
-        return $this->getValue('percentiles');
-    }
-
-    /**
-     * Get value by name.
-     *
-     * @param mixed $name
-     *
-     * @return string|array|null
-     */
-    public function getValue($name)
-    {
-        return $this->stats[$name] ?? null;
+        return $this->getStatValue($stat);
     }
 }