+++ CHANGELOG.md
@@ -7,14 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 ### Added
+- Component\Result\Stats\Result::getDistinctValues()
+- Component\Result\Stats\Result::getCountDistinct()
+- Component\Result\Stats\Result::getCardinality()
+- Component\Result\Stats\FacetValue::getPercentiles()
+- Component\Result\Stats\FacetValue::getDistinctValues()
+- Component\Result\Stats\FacetValue::getCountDistinct()
+- Component\Result\Stats\FacetValue::getCardinality()
+- Component\Result\Stats\FacetValue::getStatValue()
 
 ### Fixed
+- Component\Result\Stats\Result::getPercentiles() returns percentiles as an associative array
 
 ### Changed
+- Component\Result\Stats\Result::getMean() returns `NAN` instead of `'NaN'` if mean is NaN
+- Component\Result\Stats\FacetValue::getMean() returns `NAN` instead of `'NaN'` if mean is NaN
+- Component\Result\Stats\Result::getValue() is renamed to getStatValue()
 
 ### Removed
 
 ### Deprecated
+- Component\Result\Stats\FacetValue::getFacets()
+- Component\Result\Stats\Result::getValue()
 
 
 ## [6.1.4]
@@ -30,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Parsing nested details in debug response
 
 ### Changed
-- Solarium\Component/Result/Stats/Result::getValue() is now public
+- Solarium\Component\Result\Stats\Result::getValue() is now public
 
 
 ## [6.1.3]